Fixup tests
